refactor(singletons): move tasks to core
