+++ packages/sanity/src/_singletons/core/tasks/enabled/TasksEnabledContext.ts
@@ -1,6 +1,6 @@
 import {createContext} from 'react'
 
-import type {TasksEnabledContextValue} from '../../../core/tasks/context/enabled/types'
+import type {TasksEnabledContextValue} from '../../../../core/tasks/context/enabled/types'
 
 // NOTE: We initialize this context with a default value (`enabled: false`)
 // rather than `null` to handle cases where the tasks feature's availability
