+++ packages/sanity/src/_singletons/core/tasks/mentionUser/MentionUserContext.ts
@@ -1,6 +1,6 @@
 import {createContext} from 'react'
 
-import type {MentionUserContextValue} from '../../../core/tasks/context/mentionUser/types'
+import type {MentionUserContextValue} from '../../../../core/tasks/context/mentionUser/types'
 
 /**
  * @internal
