+++ packages/sanity/src/_singletons/core/tasks/navigation/TasksNavigationContext.ts
@@ -1,6 +1,6 @@
 import {createContext} from 'react'
 
-import type {TasksNavigationContextValue} from '../../../core/tasks/context/navigation/types'
+import type {TasksNavigationContextValue} from '../../../../core/tasks/context/navigation/types'
 
 /**
  * @internal
