+++ packages/sanity/src/_singletons/core/tasks/tasks/TasksUpsellContext.ts
@@ -1,6 +1,6 @@
 import {createContext} from 'react'
 
-import type {TasksUpsellContextValue} from '../../../core/tasks/context/upsell/types'
+import type {TasksUpsellContextValue} from '../../../../core/tasks/context/upsell/types'
 
 /**
  * @beta
