+++ packages/sanity/src/_singletons/index.ts
@@ -37,6 +37,11 @@ export * from './core/studio/router/RouterHistoryContext'
 export * from './core/studio/SourceContext'
 export * from './core/studio/WorkspaceContext'
 export * from './core/studio/workspaces/WorkspacesContext'
+export * from './core/tasks/enabled/TasksEnabledContext'
+export * from './core/tasks/mentionUser/MentionUserContext'
+export * from './core/tasks/navigation/TasksNavigationContext'
+export * from './core/tasks/tasks/TasksContext'
+export * from './core/tasks/tasks/TasksUpsellContext'
 export * from './core/user-color/UserColorManagerContext'
 export * from './router/RouterContext'
 export * from './structure/comments/authoring-path/CommentsAuthoringPathContext'
@@ -52,8 +57,3 @@ export * from './structure/components/pane/PaneLayoutContext'
 export * from './structure/components/paneRouter/PaneRouterContext'
 export * from './structure/panes/document/DocumentPaneContext'
 export * from './structure/StructureToolContext'
-export * from './tasks/enabled/TasksEnabledContext'
-export * from './tasks/mentionUser/MentionUserContext'
-export * from './tasks/navigation/TasksNavigationContext'
-export * from './tasks/tasks/TasksContext'
-export * from './tasks/tasks/TasksUpsellContext'
